feat: impl register Info
